@@ -12,7 +12,7 @@
   <url>https://github.com/matteobertozzi/dnaco-java</url>
 
   <properties>
-    <junit.version>5.7.2</junit.version>
+    <junit.version>5.8.1</junit.version>
 
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
   </properties>
@@ -40,7 +40,7 @@
         <artifactId>maven-compiler-plugin</artifactId>
         <version>3.8.1</version>
         <configuration>
-          <release>11</release>
+          <release>17</release>
           <compilerArgs>
             <arg>-parameters</arg>
             <arg>-Xlint:all</arg>
