@@ -12,10 +12,10 @@
   <url>https://github.com/matteobertozzi/dnaco-java</url>
 
   <properties>
-    <jackson.version>2.12.3</jackson.version>
-    <zstd.version>1.5.0-1</zstd.version>
+    <jackson.version>2.13.0</jackson.version>
+    <zstd.version>1.5.0-4</zstd.version>
 
-    <junit.version>5.7.2</junit.version>
+    <junit.version>5.8.1</junit.version>
 
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
   </properties>
@@ -46,11 +46,6 @@
     <dependency>
       <groupId>com.fasterxml.jackson.dataformat</groupId>
       <artifactId>jackson-dataformat-cbor</artifactId>
-      <version>${jackson.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>com.fasterxml.jackson.dataformat</groupId>
-      <artifactId>jackson-dataformat-ion</artifactId>
       <version>${jackson.version}</version>
     </dependency>
 
@@ -84,7 +79,7 @@
         <artifactId>maven-compiler-plugin</artifactId>
         <version>3.8.1</version>
         <configuration>
-          <release>11</release>
+          <release>17</release>
           <compilerArgs>
             <arg>-parameters</arg>
             <arg>-Xlint:all</arg>
