@@ -41,7 +41,7 @@ public class Histogram implements TelemetryCollector {
     1_000_000, 5000000L, 10000000L, 25000000L, 50000000L, 75000000L,          // msec
     100000000L, 150000000L, 250000000L, 350000000L, 500000000L, 750000000L,   // msec
     1000000000L, 2500000000L, 5000000000L, 10000000000L, 25000000000L,        // sec
-    50000000000L, 60000000000L, 75000000000L, 120000000000L,                  // min
+    60000000000L, 120000000000L, 300000000000L, 600000000000L, 900000000000L  // min
   };
 
   public static final long[] DEFAULT_SIZE_BOUNDS = new long[] {
